Moves `tfawserr` to `awsv1shim` package
