+++ awsv1shim/tfawserr/awserr.go
