+++ awsv1shim/tfawserr/awserr_test.go
